GameStatsService/game-scores-app/requirements.txt
uvicorn
fastapi
cassandra-driver
